refactor(traces): keep the traces plumbing off the public API
Moves resolveTracesConfig into core beside resolveMetricsConfig and
assignUserAttributes into utils, so packages/node no longer reaches for
either, and replaces the duplicated activation rule with runWithActiveSpan.
